refactor: extracted the logic in triggers into a seperate files
