+++ functions/package.json
@@ -5,7 +5,8 @@
   "author": "",
   "license": "MIT",
   "scripts": {
-    "test:firestore": "firebase emulators:exec --only firestore 'yarn jest'",
+    "test:firestore": "firebase emulators:exec --only functions,firestore 'yarn jest'",
+    "test:triggers": "firebase emulators:exec 'yarn jest ./tests/triggers/*.test.ts'",
     "test": "yarn run test:firestore",
     "lint": "eslint . --ext .ts && prettier --check \"./**/*.ts\"",
     "lint:fix": "eslint . --fix --ext .ts && prettier --write \"./**/*.ts\"",
